moved 'logprotos' to 'third_party' directory to follow current project conventions
